refactor invoice & quote
